requirements.txt
streamlit==1.56.0
faiss-cpu==1.13.2
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.0.7
google-generativeai==0.8.6
pypdf==6.9.2
python-dotenv==1.1.1
fastapi==0.136.0
uvicorn
sentence-transformers==5.4.0
pydantic
crewai==1.14.1
torch==2.11.0
rank-bm25
langgraph
langchain-groq==0.3.7
nltk 
scikit-learn